refactor: consolidate metadata loaders
